% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BostonCorr.R
\name{boston_corr}
\alias{boston_corr}
\title{Boston Housing Statistics}
\usage{
boston_corr(data)
}
\value{
the results of correlation analysis in descending order of ranking.
}
\description{
This function returns a ranking of the correlation coefficients of all variables
in the Boston Housing Dataset against 'medv'.
}
