% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BostonStat.R
\name{boston_stats}
\alias{boston_stats}
\title{Boston Housing Statistics}
\usage{
boston_stats(data, variable)
}
\value{
an object of class "boston_stats" containing the summary statistics.
}
\description{
This function returns summary statistics of any specified variable in the Boston Housing Dataset.
}
