% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BostonPlot.R
\name{BostonMEDV}
\alias{BostonMEDV}
\title{Plots MEDV against the three most significant variables.}
\usage{
BostonMEDV(data, n_vars = 3)
}
\arguments{
\item{data}{A dataframe that contains the Boston Housing Dataset.}

\item{n_vars}{The number of variables that is plotted against MEDV (number is 3).}
}
\description{
Plots MEDV against the three most significant variables.
}
